requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyswisseph==2.10.3.2
python-multipart==0.0.6
